Remove accounts from coreApi
