+++ services/core/shared/core/compat/sdk_v2/index.js
@@ -34,6 +34,8 @@ const {
 const { setCoreVersion, getCoreVersion, mapParams } = require('./coreVersionCompatibility');
 const { getConstants } = require('./constants');
 const { getBlocks } = require('./blocks');
+const { getAccounts } = require('./accounts');
+
 
 const {
 	getNetworkStatus,
@@ -142,39 +144,6 @@ const getAddressByAny = async param => {
 
 	const [prefix, body] = separateParam(param);
 	return paramNames[prefix](body);
-};
-
-const getAccounts = async params => {
-	const reqeustParams = {
-		limit: params.limit,
-		offset: params.offset,
-		sort: params.sort,
-		username: params.username,
-	};
-
-	if (params.address && typeof params.address === 'string') {
-		const parsedAddress = parseAddress(params.address);
-		if (!await confirmAddress(parsedAddress)) return {};
-		reqeustParams.address = parsedAddress;
-	}
-
-	if (params.publicKey && typeof params.publicKey === 'string') {
-		if (!validatePublicKey(params.publicKey) || (!await confirmPublicKey(params.publicKey))) {
-			return {};
-		}
-		reqeustParams.publicKey = params.publicKey;
-	}
-
-	if (params.secondPublicKey && typeof params.secondPublicKey === 'string') {
-		if (!validatePublicKey(params.secondPublicKey)
-			|| (!await confirmSecondPublicKey(params.secondPublicKey))) {
-			return {};
-		}
-		reqeustParams.secondPublicKey = params.secondPublicKey;
-	}
-
-	const result = coreApi.getAccounts(reqeustParams);
-	return result;
 };
 
 const getPublicKeyByAddress = async address => {
